+++ CHANGELOG.rst
@@ -7,6 +7,14 @@ smaht-portal
 Change Log
 ----------
 
+0.116.0
+=======
+`PR 277 DSA Release <https://github.com/smaht-dac/smaht-portal/pull/277>`_
+* In `commands/release-file.py` and `commands/create-annotated-filenames.py`:
+  * Assay and sequencer codes value set to XX for DSA fasta files and chain files
+  * For Supplementary Files, use `title` instead of `code` for software part of annotated filename
+
+
 0.115.0
 =======
 `PR 296 SN Sequencing validation <https://github.com/smaht-dac/smaht-portal/pull/296>`